DRTII-1691 Fix sdate json format. Refactor common code
