+++ src/main/scala/uk/gov/homeoffice/drt/json/SDateLikeJsonFormats.scala
@@ -1,11 +1,11 @@
 package uk.gov.homeoffice.drt.json
 
-import spray.json.{DefaultJsonProtocol, JsNumber, JsValue, RootJsonFormat, enrichAny}
+import spray.json.{DefaultJsonProtocol, JsNumber, JsString, JsValue, RootJsonFormat, enrichAny}
 import uk.gov.homeoffice.drt.time.{SDate, SDateLike}
 
 object SDateLikeJsonFormats extends DefaultJsonProtocol {
 
-  implicit object SDateLikeJsonFormat extends RootJsonFormat[SDateLike] {
+  implicit object SDateLikeTimestampJsonFormat extends RootJsonFormat[SDateLike] {
     override def read(json: JsValue): SDateLike = json match {
       case JsNumber(ts) => SDate(ts.toLong)
     }
