+++ src/main/scala/uk/gov/homeoffice/drt/services/api/v1/serialiser/QueueApiV1JsonFormats.scala
@@ -1,15 +1,11 @@
 package uk.gov.homeoffice.drt.services.api.v1.serialiser
 
 import spray.json._
-import uk.gov.homeoffice.drt.json.SDateLikeJsonFormats.SDateLikeJsonFormat
-import uk.gov.homeoffice.drt.ports.PortCode
 import uk.gov.homeoffice.drt.ports.Queues.Queue
-import uk.gov.homeoffice.drt.ports.Terminals.Terminal
 import uk.gov.homeoffice.drt.routes.api.v1.QueueApiV1Routes.QueueJsonResponse
 import uk.gov.homeoffice.drt.services.api.v1.QueueExport.{PeriodJson, PortQueuesJson, QueueJson, TerminalQueuesJson}
 
-trait QueueApiV1JsonFormats extends DefaultJsonProtocol {
-
+trait QueueApiV1JsonFormats extends DefaultJsonProtocol with CommonJsonFormats {
   implicit object QueueJsonFormat extends RootJsonFormat[Queue] {
     override def write(obj: Queue): JsValue = obj.stringValue.toJson
 
@@ -23,25 +19,7 @@ trait QueueApiV1JsonFormats extends DefaultJsonProtocol {
 
   implicit val periodJsonFormat: RootJsonFormat[PeriodJson] = jsonFormat2(PeriodJson.apply)
 
-  implicit object TerminalJsonFormat extends RootJsonFormat[Terminal] {
-    override def write(obj: Terminal): JsValue = obj.toString.toJson
-
-    override def read(json: JsValue): Terminal = json match {
-      case JsString(value) => Terminal(value)
-      case unexpected => throw new Exception(s"Failed to parse Terminal. Expected JsString. Got ${unexpected.getClass}")
-    }
-  }
-
   implicit val terminalQueuesJsonFormat: RootJsonFormat[TerminalQueuesJson] = jsonFormat2(TerminalQueuesJson.apply)
-
-  implicit object PortCodeJsonFormat extends RootJsonFormat[PortCode] {
-    override def write(obj: PortCode): JsValue = obj.iata.toJson
-
-    override def read(json: JsValue): PortCode = json match {
-      case JsString(value) => PortCode(value)
-      case unexpected => throw new Exception(s"Failed to parse Terminal. Expected JsString. Got ${unexpected.getClass}")
-    }
-  }
 
   implicit val portQueuesJsonFormat: RootJsonFormat[PortQueuesJson] = jsonFormat2(PortQueuesJson.apply)
